chore(suite-native): Move useTradingFiatValues to common
